requirements.txt
amqp==5.0.6
asgiref==3.3.1
billiard==3.6.4.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-ckeditor==6.0.0
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.0.2
MarkupSafe==2.0.1
mysqlclient==2.0.3
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
prompt-toolkit==3.0.18
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
